requirements.txt
fastapi[standard]
motor
pymongo
aiofiles
pydantic>=2.0.0
python-jose[cryptography] 
python-dotenv 
httpx
fastapi_oidc
paramiko
     


